Rename 'client_exporter' folder to 'exporter'
