+++ js/exporter/excel_creator.js
@@ -7,7 +7,7 @@ var Class = require("../core/class"),
     JSZip = require("jszip"),
     fileSaver = require("./file_saver"),
     excelFormatConverter = require("./excel_format_converter"),
-    XlsxFile = require("./xlsx/xlsx_file"),
+    XlsxFile = require("./excel/excel.file"),
     isDefined = typeUtils.isDefined,
     XML_TAG = "<?xml version=\"1.0\" encoding=\"utf-8\"?>",
     GROUP_SHEET_PR_XML = "<sheetPr><outlinePr summaryBelow=\"0\"/></sheetPr>",
