+++ testing/tests/DevExpress.ui.widgets.pivotGrid/pivotGrid.export.tests.js
@@ -7,8 +7,8 @@ import "ui/pivot_grid/ui.pivot_grid";
 
 import $ from "jquery";
 import { excel as excelCreator } from "client_exporter";
-import { __internals as internals } from "client_exporter/excel_creator";
-import excel_creator from "client_exporter/excel_creator";
+import { __internals as internals } from "exporter/excel_creator";
+import excel_creator from "exporter/excel_creator";
 import JSZipMock from "../../helpers/jszipMock.js";
 
 const BASE_STYLE_XML1 = "<fonts count=\"2\"><font><sz val=\"11\" /><color theme=\"1\" /><name val=\"Calibri\" /><family val=\"2\" />" +
